pyproject.toml
[build-system]
requires = ['setuptools>=61.0']
build-backend = 'setuptools.build_meta'

[project]
name = 'news-feed-api'
version = '1.0.0'
description = 'API for news feed aggregation'
requires-python = ">=3.10"
dependencies = [
  "aioredis==2.0.1",
  "alembic==1.14.1",
  "amqp==5.3.1",
  "annotated-types==0.7.0",
  "anyio==4.8.0",
  "async-timeout==5.0.1",
  "bcrypt==4.2.1",
  "billiard==4.2.1",
  "celery==5.4.0",
  "certifi==2025.1.31",
  "cffi==1.17.1",
  "click==8.1.8",
  "click-didyoumean==0.3.1",
  "click-plugins==1.1.1",
  "click-repl==0.3.0",
  "cryptography==44.0.1",
  "ecdsa==0.19.0",
  "fastapi==0.115.8",
  "greenlet==3.1.1",
  "h11==0.14.0",
  "httpcore==1.0.7",
  "httpx==0.28.1",
  "idna==3.10",
  "kombu==5.4.2",
  "Mako==1.3.9",
  "MarkupSafe==3.0.2",
  "passlib==1.7.4",
  "prompt_toolkit==3.0.50",
  "pyasn1==0.4.8",
  "pycparser==2.22",
  "pydantic==2.10.6",
  "pydantic-settings==2.7.1",
  "pydantic_core==2.27.2",
  "python-dateutil==2.9.0.post0",
  "python-dotenv==1.0.1",
  "python-jose==3.4.0",
  "redis==5.2.1",
  "rsa==4.9",
  "six==1.17.0",
  "sniffio==1.3.1",
  "SQLAlchemy==2.0.38",
  "sqlmodel==0.0.22",
  "starlette==0.45.3",
  "typing_extensions==4.12.2",
  "tzdata==2025.1",
  "uvicorn==0.34.0",
  "vine==5.1.0",
  "wcwidth==0.2.13"]

[project.optional-dependencies]
dev = [
  "factory_boy==3.3.3",
  "Faker==36.1.1",
  "iniconfig==2.0.0",
  "packaging==24.2",
  "pluggy==1.5.0",
  "pytest==8.3.4",
  'pytest-asyncio>=0.21.0',
  'pytest-cov>=4.1.0',
]
[tools.pytest.ini_options]
test_paths = ['tests']
python_files = 'test_*.py'
python_classes = 'Test*'
python_functions = 'test_*'
asyncio_mode = 'auto'

[tool.setuptools]
package-dir = {"" = "."}
packages = ['src']
